mini correcciones
se estilizó un poco el home y se corrigieron algunas cosas con respecto a los nombres de algunos componentes
