+++ ProyectFinalGrupo2/public/styles/Style.css
@@ -6,32 +6,9 @@ body {
     background-size: cover;
 }
 
-h1, h2, h3, h4, h5, h6 {
-    color: #ffffff;
-    margin-bottom: 1rem;
-}
-/*
-.navbar {
-    background-color: #2a2a2a;
-}
-
-.navbar .nav-link {
-    color: #f5f5f5;
-    font-weight: 600;
-    transition: background-color 0.3s ease;
-}
-
-.navbar .nav-link:hover {
-    background-color: #4a4a4a;
-}
-
-.navbar-brand img {
-    max-height: 60px;
-}
-*/
 .section {
     padding: 3rem 1.5rem;
-   /* background-color: #fff;*/
+
 }
 
 .principal img {
@@ -40,6 +17,7 @@ h1, h2, h3, h4, h5, h6 {
 }
 
 .card {
+    background: linear-gradient(0deg, rgb(108, 134, 208) 0%, #40337c 100%);
     border-radius: 8px;
     box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
     margin-bottom: 2rem;
@@ -58,7 +36,9 @@ h1, h2, h3, h4, h5, h6 {
     font-size: 1rem;
     margin-bottom: 1rem;
 }
-
+.listaIntegrantes{
+    color: #ffffff;
+}
 .image img {
     border-radius: 8px;
     transition: transform 0.3s ease;
@@ -67,23 +47,9 @@ h1, h2, h3, h4, h5, h6 {
 .image img:hover {
     transform: scale(1.05);
 }
-/*
-.footer {
-    padding: 2rem;
-    background-color: #2a2a2a;
-}
 
-.footer p {
-    color: #f5f5f5;
-    margin: 0;
-}
-*/
 /* Responsividad */
-@media (max-width: 768px) {
-    .navbar {
-        background-color: #2a2a2a;
-    }
-    
+@media (max-width: 768px) { 
     .section {
         padding: 1.5rem 1rem;
     }
