+++ ProyectFinalGrupo2/src/App.jsx
@@ -7,7 +7,6 @@ import AboutUs from './Components/AboutUs';
 import CalculadoraIMC from './Components/CalculadoraIMC';
 import GestorBilleteraVirtual from './Components/GestorBilleteraVirtual';
 import JuegoPhaser from './Components/JuegoPhaser';
-import IMCCalcular from './Components/IMCCalcular';
 import JuegoReact from './Components/desafioMatematico/JuegoReact';
 import ErrorPage from './Components/ErrorPage';
 
@@ -23,7 +22,6 @@ function App() {
         <Route path='calculadoraIMC' element={ <CalculadoraIMC/> } />
         <Route path='gestorBilleteraVirtual' element={ <GestorBilleteraVirtual/> } />
         <Route path='juegoPhaser' element={ <JuegoPhaser/> } />
-        <Route path='IMCCalcular' element={<IMCCalcular />} />
         <Route path='juegoReact' element={ <JuegoReact/> } />
         <Route path='*' element={ <ErrorPage/> } />    
       </Route>
