+++ ProyectFinalGrupo2/src/Components/Layout.jsx
@@ -17,8 +17,6 @@
                             <Nav className="me-auto">
                                 <Nav.Link href="/">Home</Nav.Link>
                                 <Nav.Link href="/aboutUs">AboutUs</Nav.Link>
-                                <Nav.Link href="/juegoPhaser">JuegoPhaser</Nav.Link>
-                                <Nav.Link href ="/IMCCalcular">Salud Jugador:IMC</Nav.Link>
                                 <NavDropdown
                                  id="nav-dropdown-dark-example"
                                  title="Proyectos"
